+++ cmd/neofs-adm/internal/modules/config/config.go
@@ -42,8 +42,7 @@ network:
     container_alias: {{ .ContainerAliasFee }}
     withdraw: {{ .WithdrawFee}}
 # if credentials section is omitted, then neofs-adm will require manual password input
-credentials:
-  contract: password # wallet for contract group signature{{ range.Glagolitics}}
+credentials:{{ range.Glagolitics}}
   {{.}}: password{{end}}
 `