+++ cmd/neofs-adm/internal/modules/morph/initialize_register.go
@@ -9,6 +9,8 @@ import (
 	"github.com/nspcc-dev/neo-go/pkg/core/transaction"
 	"github.com/nspcc-dev/neo-go/pkg/io"
 	"github.com/nspcc-dev/neo-go/pkg/rpcclient"
+	"github.com/nspcc-dev/neo-go/pkg/rpcclient/actor"
+	"github.com/nspcc-dev/neo-go/pkg/rpcclient/invoker"
 	"github.com/nspcc-dev/neo-go/pkg/rpcclient/neo"
 	"github.com/nspcc-dev/neo-go/pkg/rpcclient/unwrap"
 	"github.com/nspcc-dev/neo-go/pkg/smartcontract/callflag"
@@ -49,12 +51,15 @@ func (c *initializeContext) registerCandidates() error {
 		panic(fmt.Sprintf("BUG: %v", w.Err))
 	}
 
-	signers := []rpcclient.SignerAccount{{
-		Signer:  c.getSigner(false, c.CommitteeAcc),
+	signers := []actor.SignerAccount{{
+		Signer: transaction.Signer{
+			Account: c.CommitteeAcc.Contract.ScriptHash(),
+			Scopes:  transaction.CalledByEntry,
+		},
 		Account: c.CommitteeAcc,
 	}}
 	for i := range c.Accounts {
-		signers = append(signers, rpcclient.SignerAccount{
+		signers = append(signers, actor.SignerAccount{
 			Signer: transaction.Signer{
 				Account:          c.Accounts[i].Contract.ScriptHash(),
 				Scopes:           transaction.CustomContracts,
@@ -63,8 +68,12 @@ func (c *initializeContext) registerCandidates() error {
 			Account: c.Accounts[i],
 		})
 	}
+	act, err := actor.New(c.Client, signers)
+	if err != nil {
+		return fmt.Errorf("creating actor: %w", err)
+	}
 
-	tx, err := c.Client.CreateTxFromScript(w.Bytes(), c.CommitteeAcc, -1, 0, signers)
+	tx, err := act.MakeUnsignedRun(w.Bytes(), nil)
 	if err != nil {
 		return fmt.Errorf("can't create tx: %w", err)
 	}
@@ -83,9 +92,7 @@ func (c *initializeContext) registerCandidates() error {
 }
 
 func (c *initializeContext) transferNEOToAlphabetContracts() error {
-	neoHash := neo.Hash
-
-	ok, err := c.transferNEOFinished(neoHash)
+	ok, err := c.transferNEOFinished()
 	if ok || err != nil {
 		return err
 	}
@@ -96,7 +103,7 @@ func (c *initializeContext) transferNEOToAlphabetContracts() error {
 	bw := io.NewBufBinWriter()
 	for _, acc := range c.Accounts {
 		h := state.CreateContractHash(acc.Contract.ScriptHash(), cs.NEF.Checksum, cs.Manifest.Name)
-		emit.AppCall(bw.BinWriter, neoHash, "transfer", callflag.All,
+		emit.AppCall(bw.BinWriter, neo.Hash, "transfer", callflag.All,
 			c.CommitteeAcc.Contract.ScriptHash(), h, int64(amount), nil)
 		emit.Opcodes(bw.BinWriter, opcode.ASSERT)
 	}
@@ -108,9 +115,10 @@ func (c *initializeContext) transferNEOToAlphabetContracts() error {
 	return c.awaitTx()
 }
 
-func (c *initializeContext) transferNEOFinished(neoHash util.Uint160) (bool, error) {
-	bal, err := c.Client.NEP17BalanceOf(neoHash, c.CommitteeAcc.Contract.ScriptHash())
-	return bal < native.NEOTotalSupply, err
+func (c *initializeContext) transferNEOFinished() (bool, error) {
+	neoR := neo.NewReader(invoker.New(c.Client, nil))
+	bal, err := neoR.BalanceOf(c.CommitteeAcc.Contract.ScriptHash())
+	return bal.Int64() < native.NEOTotalSupply, err
 }
 
 var errGetPriceInvalid = errors.New("`getRegisterPrice`: invalid response")