+++ pkg/innerring/innerring.go
@@ -10,7 +10,6 @@ import (
 	"sync/atomic"
 
 	"github.com/nspcc-dev/neo-go/pkg/core/block"
-	"github.com/nspcc-dev/neo-go/pkg/core/transaction"
 	"github.com/nspcc-dev/neo-go/pkg/crypto/keys"
 	"github.com/nspcc-dev/neo-go/pkg/encoding/fixedn"
 	"github.com/nspcc-dev/neofs-node/misc"
@@ -135,7 +134,6 @@ type (
 		cfg  *viper.Viper
 		key  *keys.PrivateKey
 		name string
-		sgn  *transaction.Signer
 		from uint32 // block height
 	}
 )
@@ -417,6 +415,7 @@ func New(ctx context.Context, log *zap.Logger, cfg *viper.Viper, errChan chan<- 
 			client.WithContext(ctx),
 			client.WithLogger(log),
 			client.WithSingleClient(wsClient),
+			client.WithAutoSidechainScope(),
 		)
 		if err != nil {
 			return nil, fmt.Errorf("init internal morph client: %w", err)
@@ -450,9 +449,6 @@ func New(ctx context.Context, log *zap.Logger, cfg *viper.Viper, errChan chan<- 
 	if err != nil {
 		return nil, err
 	}
-	if err := server.morphClient.SetGroupSignerScope(); err != nil {
-		morphChain.log.Info("failed to set group signer scope, continue with Global", zap.Error(err))
-	}
 
 	server.withoutMainNet = cfg.GetBool("without_mainnet")
 
@@ -465,7 +461,6 @@ func New(ctx context.Context, log *zap.Logger, cfg *viper.Viper, errChan chan<- 
 	} else {
 		mainnetChain := morphChain
 		mainnetChain.name = mainnetPrefix
-		mainnetChain.sgn = &transaction.Signer{Scopes: transaction.CalledByEntry}
 
 		fromMainChainBlock, err := server.persistate.UInt32(persistateMainChainLastBlockKey)
 		if err != nil {
@@ -988,16 +983,13 @@ func (s *Server) createClient(ctx context.Context, p chainParams, errChan chan<-
 	if len(endpoints) == 0 {
 		return nil, fmt.Errorf("%s chain client endpoints not provided", p.name)
 	}
-
-	return client.New(
-		p.key,
+	var options = []client.Option{
 		client.WithContext(ctx),
 		client.WithLogger(p.log),
-		client.WithDialTimeout(p.cfg.GetDuration(p.name+".dial_timeout")),
-		client.WithSigner(p.sgn),
+		client.WithDialTimeout(p.cfg.GetDuration(p.name + ".dial_timeout")),
 		client.WithEndpoints(endpoints),
-		client.WithReconnectionRetries(p.cfg.GetInt(p.name+".reconnections_number")),
-		client.WithReconnectionsDelay(p.cfg.GetDuration(p.name+".reconnections_delay")),
+		client.WithReconnectionRetries(p.cfg.GetInt(p.name + ".reconnections_number")),
+		client.WithReconnectionsDelay(p.cfg.GetDuration(p.name + ".reconnections_delay")),
 		client.WithConnSwitchCallback(func() {
 			var err error
 
@@ -1014,7 +1006,12 @@ func (s *Server) createClient(ctx context.Context, p chainParams, errChan chan<-
 			errChan <- fmt.Errorf("%s chain connection has been lost", p.name)
 		}),
 		client.WithMinRequiredBlockHeight(p.from),
-	)
+	}
+	if p.name == morphPrefix {
+		options = append(options, client.WithAutoSidechainScope())
+	}
+
+	return client.New(p.key, options...)
 }
 
 const validatorsConfigKey = "morph.validators"