+++ CHANGELOG.md
@@ -19,11 +19,13 @@ Changelog for NeoFS Node
 - FSTree storage now uses more efficient and safe temporary files under Linux (#2566)
 - BoltDB open timeout increased from 100ms to 1s (#2499)
 - Internal container cache size from 10 to 1000 (#2600)
+- Transaction witness scope no longer uses CustomGroups relying on more strict Rules now (#2619)
 
 ### Removed
 - deprecated `no-precheck` flag of `neofs-cli container set-eacl` (#2496)
 - Recently-handled objects Policer's cache (#2600)
 - deprecated blobovnicza substorage (#2614)
+- Contract group wallet support from `neofs-adm` (#2619)
 
 ### Updated
 - Update minimal supported Go version up to v1.19 (#2485)
@@ -37,6 +39,9 @@ Changelog for NeoFS Node
 Blobovniczas are gone from the node with this release, see 0.38.0 upgrade
 notes for migration instruction (the tool is still provided, but will be gone
 after a couple of minor releases).
+
+neofs-adm no longer creates a contract group wallet and no longer needs it
+for the operation which may affect deployment scripts.
 
 ## [0.38.1] - 2023-09-18